backend/pyproject.toml
[project]
name = "advanced-rag"
version = "0.1.0"
description = "Advanced Hybrid RAG backend with shared core runtime for API and MCP"
requires-python = ">=3.13"
dependencies = [
    # Core framework
    "fastapi>=0.115.0",
    "fastapi-offline>=1.7.0",
    "uvicorn[standard]>=0.32.0",
    "python-multipart>=0.0.12",
    # FastMCP server for exposing the query capability to MCP clients
    "fastmcp>=3.0.0",
    # Haystack 2.x  —  2.25.1 is the current stable release (Feb 2026)
    # SentenceTransformersSimilarityRanker (replaces deprecated TransformersSimilarityRanker)
    # was introduced in 2.9.  HierarchicalDocumentSplitter available since 2.7.
    "haystack-ai>=2.25.1",
    # Qdrant integration (dense + sparse named-vector collection)
    "qdrant-haystack>=10.2.0",
    # Sparse embeddings: SPLADE / BM42 via FastEmbed (local ONNX, no API)
    "fastembed-haystack>=2.1.0",
    # Pydantic settings
    "pydantic-settings>=2.0",
    "pydantic>=2.0",
    # HTTP client for docling-serve v1 API
    "httpx>=0.27.0",
    # HuggingFace — dense embeddings + cross-encoder reranker (local inference)
    "sentence-transformers>=3.0.0",
    "transformers>=4.40.0",
    "torch>=2.0.0",
    # OpenAI SDK — for OpenAI-compatible LLM endpoints
    # (OpenAI, Ollama, vLLM, LM Studio, Groq, …)
    "openai>=1.0.0",
    # Markdown token parser for heading and structure-aware metadata
    "markdown-it-py>=4.0.0",
    # Deterministic document date extraction from titles / openings
    "dateparser>=1.2.1",
    # ColBERT late-interaction reranker (local HuggingFace, no index change needed)
    "pylate>=1.1.0",
    # MinIO — object storage client for original document files
    "minio>=7.0.0",
    # RAGAS evaluation framework (lazy-imported, only when RAGAS_ENABLED=true)
    "ragas>=0.2.0",
    "datasets>=2.0.0",
    "nltk>=3.9.1",
]

[project.scripts]
advanced-rag-api = "adapters.api.app:main"
advanced-rag-mcp = "adapters.mcp.server:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

# Tell hatchling which sub-packages to include (this is an app, not a library)
[tool.hatch.build.targets.wheel]
packages = ["adapters", "core"]

[dependency-groups]
dev = [
    "httpx",   # for FastAPI TestClient
    "ruff",
]

[tool.ruff]
line-length = 120

[tool.ruff.lint]
select = ["E", "F", "I"]  # pycodestyle + pyflakes + isort

[tool.ruff.lint.per-file-ignores]
# LLM prompt strings contain intentionally long few-shot examples
"core/components/query_analyzer.py" = ["E501"]
